1. PDF Summaries/requirements.txt
python-dotenv
openai
PyPDF2
backoff 